Add Programatic solution to problem 10 based off of problem 7
Use the optimezed O(1) space prime number generator and the efficient
prime number candidate loop to generate a sum instead of the largest
prime number under the limit. Both the iterator and the prime number
checker are a very optimized O(n). The O(N) space version may be slightly
faster but that isn't a major concern atm