Add object-based solution to E11 based off my algorithm for E8
My solution this time uses a "FactorQueue" to group all the
datatypes I was setting on each pass. in addition it stores the max
value it has passed over. This solves the problem in O(d) space
where d is the number of consecutive digits we are looking for
the product of. it also only touches each element of the grid at
most 4d times so runtime complexity is O(dn^2) where n is the length
of the 2-dimensional grid. I also opted not to use one of my breaking
conditions because, when added to each loop, It became a lot less
legible. Possible optimizations include:
    *Using multiple factorQueues to parallelize work
    *Reintroducing the early break for when we find a zero with less than d digits left in the line
    *Reintroducing the breaking condition when finding the maximum possible value before fully traversing